FROM LEGION 5iPRO

+++ src/binarySearch/medium/MinimizeMaxDistanceBetweenGasStations/BruteForce.java
@@ -2,19 +2,19 @@ package binarySearch.medium.MinimizeMaxDistanceBetweenGasStations;
 
 public class BruteForce
 {
-    private static int findMinimizedMax(int[] nums, int k)
+    private static double findMinimizedMax(int[] nums, int k)
     {
         int[] placed = new int[nums.length - 1];
 
         for(int gs = 1 ; gs <= k ; gs++)
         {
-            int maxSection = -1;
+            double maxSection = -1;
             int maxIndex = -1;
 
             for(int i = 0 ; i < nums.length - 1 ; i++)
             {
-                int diff = nums[i + 1] - nums[i];
-                int sectionLength = diff / (placed[i] + 1);
+                double diff = nums[i + 1] - nums[i];
+                double sectionLength = diff / (double)(placed[i] + 1);
 
                 if(sectionLength > maxSection)
                 {
@@ -25,12 +25,12 @@ public class BruteForce
             placed[maxIndex]++;
         }
 
-        int maxAns = -1;
+        double maxAns = -1;
 
         for(int i = 0 ; i < nums.length - 1 ; i++)
         {
-            int diff = nums[i + 1] - nums[i];
-            int sectionLength = diff / (placed[i] + 1);
+            double diff = nums[i + 1] - nums[i];
+            double sectionLength = diff / (double)(placed[i] + 1);
             maxAns = Math.max(maxAns, sectionLength);
         }
 
@@ -42,7 +42,7 @@ public class BruteForce
         int[] nums = {1, 13, 17, 23};
         int k = 5;
 
-        int minimizedMaxDistance = findMinimizedMax(nums, k);
+        double minimizedMaxDistance = findMinimizedMax(nums, k);
 
         System.out.print(minimizedMaxDistance);
     }
